turn ArticleService param contributor_id to token,add ContributorController

+++ application/literatureClub/src/main/java/com/snach/literatureclub/service/ArticleService.java
@@ -2,7 +2,7 @@ package com.snach.literatureclub.service;
 
 import com.snach.literatureclub.bean.Article;
 import com.snach.literatureclub.dao.ArticleDao;
-import com.snach.literatureclub.utils.Tools;
+import com.snach.literatureclub.utils.IdTools;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,54 +11,62 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.snach.literatureclub.utils.Tools.generateId;
+import static com.snach.literatureclub.utils.TokenTools.getPayload;
+import static com.snach.literatureclub.utils.TokenTools.tokenVerify;
+import static com.snach.literatureclub.utils.IdTools.generateId;
 
 @Service
 public interface ArticleService {
     /**
      * 用户新增稿件，id由时间戳生成，将稿件根据action进行保存
      * 若action为草稿保存则无需审核，为发布则需要审核
-     * 返回各事件执行状态
+     * 返回稿件基本信息（标题、描述、时间和id）和各事件执行状态
+     * 返回格式{article_id: #{String}, title: #{String}, description: #{String}, time:#{Date}, fileStatue: #{INTEGER}, statusMsg: #{STRING} }
      *
-     * @param contributor_id
-     * @param title          标题
-     * @param description    描述
-     * @param text           稿件内容
-     * @param action         稿件处理事件（1：草稿保存 2：发布 ）
-     * @return 保存状态（1：保存成功 2：待审核 0：保存失败）,执行状态 返回格式{ fileStatue: #{INTEGER}, statusMsg: #{STRING} }
+     * @param token       用于验证是否过期以及获取作者id
+     * @param article_id  稿件id，初次创建为null
+     * @param title       标题
+     * @param description 描述
+     * @param text        稿件内容
+     * @param action      稿件处理事件（1：草稿保存 2：发布 ）
+     * @return 稿件基本信息（标题、描述、时间和id）,保存状态（1：保存成功 2：待审核 0：保存失败）,执行状态
      */
-    Map<String, Object> addArticle(String contributor_id, String title, String description, String text, int action);
+    Map<String, Object> addArticle(String token, String article_id, String title, String description, String text, int action);
 
     /**
      * 用户更改稿件基础信息，包括标题和描述，根据id进行更新，同时更新修改时间
-     *
+     * 返回格式{article_id: #{String}, title: #{String}, description: #{String}, time:#{Date}, fileStatue: #{INTEGER}, statusMsg: #{STRING} }
+     * @param token       用于验证是否过期以及获取作者id
      * @param id          稿件id
      * @param title       标题
      * @param description 描述
      * @return 保存状态（2：待审核 0：保存失败）,执行状态 返回格式{ fileStatue: #{INTEGER}, statusMsg: #{STRING} }
      */
-    Map<String, Object> updateArticle(String id, String title, String description);
+    Map<String, Object> updateArticle(String token, String id, String title, String description);
 
     /**
      * 用户更改稿件详细信息，包括标题、描述和内容，根据id进行更新，同时更新修改时间
      * 若action为草稿保存则无需审核，为发布则需要审核
+     * 返回格式{article_id: #{String}, title: #{String}, description: #{String}, time:#{Date}, fileStatue: #{INTEGER}, statusMsg: #{STRING} }
      *
+     * @param token
      * @param id          稿件id
      * @param title       标题
      * @param description 描述
      * @param text        内容
      * @param action      稿件处理事件（1：草稿保存 2：发布 ）
-     * @return 保存状态（1：保存成功 2：待审核 0：保存失败）,执行状态 返回格式{ fileStatue: #{INTEGER}, statusMsg: #{STRING} }
+     * @return 保存状态（1：保存成功 2：待审核 0：保存失败）,执行状态
      */
-    Map<String, Object> updateArticle(String id, String title, String description, String text, int action);
+    Map<String, Object> updateArticle(String token,String id, String title, String description, String text, int action);
 
     /**
      * 根据id删除稿件
      *
-     * @param id 稿件id
+     * @param token
+     * @param id    稿件id
      * @return 执行状态 返回格式{ statusMsg: #{STRING} }
      */
-    Map<String, Object> deleteArticle(String id);
+    Map<String, Object> deleteArticle(String token, String id);
 
     /**
      * 根据作者id查找其稿件基础信息，包括标题、修改时间和描述，用于创作者界面显示
@@ -75,6 +83,14 @@ public interface ArticleService {
      * @return 稿件详细信息, 执行状态 返回格式{ article: #{ARTICLE}, statusMsg: #{STRING}}
      */
     Map<String, Object> getArticleById(String id);
+
+    /**
+     * 返回所有稿件的详细信息
+     * 返回格式 { articles: [#{ARTICLE},...], statusMsg: #{STRING}}
+     *
+     * @return 所有的稿件信息的Article对象List
+     */
+    Map<String, Object> getAllArticles();
 }
 
 @Service
@@ -84,69 +100,114 @@ class ArticleServiceImpl implements ArticleService {
     private ArticleDao articleDao;
 
     @Override
-    public Map<String, Object> addArticle(String contributor_id, String title, String description, String text, int action) {
+    public Map<String, Object> addArticle(String token, String article_id, String title, String description, String text, int action) {
         Map<String, Object> res = new HashMap<String, Object>();
-        //时间戳生成id
-        String id = generateId(Tools.Type.ARTICLE);
+        // 检测token是否合法
+        if (!tokenVerify(token)) {
+            res.put("statusMsg", "Invalid token.");
+            return res;
+        }
+        // 获取作者id
+        String contributor_id = getPayload(token,"id");
+        if (article_id == null) {
+            //没有稿件id，时间戳生成id
+            article_id = generateId(IdTools.Type.ARTICLE);
+        } else {
+            //若已有稿件id，则进行更新
+            return updateArticle(token, article_id, title, description, text, action);
+        }
 
         //修改时间
         //SimpleDateFormat formatter= new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
         Date date = new Date(System.currentTimeMillis());
 
         //插入article表
-        Article article = new Article(id, text, date, "", title, description,action);
+        Article article = new Article(article_id, text, date, "", title, description, action);
         articleDao.insertArticle(article);
         // 稿件投稿者关系更新
-        articleDao.insertRelationAndState(contributor_id, id);
+        articleDao.insertRelation(contributor_id, article_id);
 
+        res.put("article_id", article_id);
+        res.put("title", title);
+        res.put("description", description);
+        res.put("time", date);
         res.put("fileStatue", action);
         res.put("statusMsg", "ok");
         return res;
     }
 
     @Override
-    public Map<String, Object> updateArticle(String id, String title, String description) {
+    public Map<String, Object> updateArticle(String token,String id, String title, String description) {
         Map<String, Object> res = new HashMap<String, Object>();
         // 修改时间
         Date date = new Date(System.currentTimeMillis());
 
         // 更新基本信息
-        Article article = new Article(id, "", date, "", title, description,2);
+        Article article = new Article(id, "", date, "", title, description, 2);
         articleDao.updateArticleInfo(article);
 
+        res.put("article_id", id);
+        res.put("title", title);
+        res.put("description", description);
+        res.put("time", date);
         res.put("fileStatue", 2);
         res.put("statusMsg", "ok");
         return res;
     }
 
     @Override
-    public Map<String, Object> updateArticle(String id, String title, String description, String text, int action) {
+    public Map<String, Object> updateArticle(String token, String id, String title, String description, String text, int action) {
         Map<String, Object> res = new HashMap<String, Object>();
-
+        // 检测token是否合法
+        if (!tokenVerify(token)) {
+            res.put("statusMsg", "Invalid token.");
+            return res;
+        }
+        // 获取作者id
+        String contributor_id = getPayload(token,"id");
+        if(articleDao.belong(contributor_id,id)==0){
+            res.put("statusMsg", "Access denied");
+            return res;
+        }
         // 修改时间
         Date date = new Date(System.currentTimeMillis());
         // 更新详细信息
-        Article article = new Article(id, text, date, "", title, description,action);
+        Article article = new Article(id, text, date, "", title, description, action);
         articleDao.updateArticleDetail(article);
 
+        res.put("article_id", id);
+        res.put("title", title);
+        res.put("description", description);
+        res.put("time", date);
         res.put("fileStatue", action);
         res.put("statusMsg", "ok");
         return res;
     }
 
     @Override
-    public Map<String, Object> deleteArticle(String id) {
+    public Map<String, Object> deleteArticle(String token, String id) {
         Map<String, Object> res = new HashMap<String, Object>();
+        // 检测token是否合法
+        if (!tokenVerify(token)) {
+            res.put("statusMsg", "Invalid token.");
+            return res;
+        }
+        // 获取作者id
+        String contributor_id = getPayload(token,"id");
+        if(articleDao.belong(contributor_id,id)==0){
+            res.put("statusMsg", "Access denied");
+            return res;
+        }
         articleDao.deleteArticleById(id);
-        res.put("statusMsg", "ok");
+        res.put("statusMsg", "success");
         return res;
     }
 
     @Override
     public Map<String, Object> getContributorArticles(String contributor_id) {
         Map<String, Object> res = new HashMap<String, Object>();
-        res.put("articles",articleDao.getArticleByContributorId(contributor_id));
-        res.put("statusMsg", "ok");
+        res.put("articles", articleDao.getArticleByContributorId(contributor_id));
+        res.put("statusMsg", "success");
         return res;
     }
 
@@ -154,7 +215,15 @@ class ArticleServiceImpl implements ArticleService {
     public Map<String, Object> getArticleById(String id) {
         Map<String, Object> res = new HashMap<String, Object>();
         res.put("article", articleDao.getArticleById(id));
-        res.put("statusMsg", "ok");
+        res.put("statusMsg", "success");
+        return res;
+    }
+
+    @Override
+    public Map<String, Object> getAllArticles() {
+        Map<String, Object> res = new HashMap<String, Object>();
+        res.put("article", articleDao.getAllArticles());
+        res.put("statusMsg", "success");
         return res;
     }
 }
